% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eyam.R
\docType{data}
\name{Eyam}
\alias{Eyam}
\title{Eyam plague.}
\format{A data frame with 8 rows and 4 variables:
\describe{
 \item{time}{Months past June 18 1666}
 \item{S}{Susceptible}
 \item{I}{Infectious}
 \item{R}{Removed} 
}}
\usage{
data(Eyam)
}
\description{
A dataset containing the number of susceptible, infectious and removed
individuals during the Eyam plague from June 18 to October 20, 1666.
}
\references{
Ragget G (1982). A stochastic model of the Eyam plague. Journal of Applied Statistics 9, 212-226.
}

